requirements.txt
groq
SpeechRecoginition
pyttsx3
pyaudio
python-dotenv
streamlit
